Fix integer dtype

+++ tests/dsp/multirate/test_decimator.py
@@ -47,7 +47,7 @@ def test_streaming_match_full():
     N = 360
     rng = np.random.default_rng()
     x = rng.standard_normal(N) + 1j * rng.standard_normal(N)  # Input signal
-    r = rng.integers(3, 7)  # Interpolation rate
+    r = rng.integers(3, 7, dtype=int)  # Interpolation rate
 
     fir1 = sdr.Decimator(r)
     y_full = fir1(x, mode="full")
